Move mode helpers into submodule
