+++ src/core/mode/minified-mode.js
